Submit of hw 6, forms
Hey! I left a note in the js about this, but here’s what I was
wondering: for the phone number input field, I made an alert in js to
prompt user to enter phone number correctly, by length. I’m wondering 2
things: 1) do non-numerical characters like “( )” and “-“ count for the
maxlength, i.e., if the phone number is 703 854 9099, that is 10
characters, and with (703) 854-9099, would this be 13 characters? 2)
I’ve seen in some forms I’ve used before something that looks like js
to automatically add the “( )” and “-“ in formatting a phone number as
it is entered. Couldn’t find an answer in googleying around! Just
curious :)